# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Tencent_WeKnora_internal_errors.AppError:
    properties:
      code:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_Tencent_WeKnora_internal_errors.ErrorCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 2000
    - 2001
    - 2002
    - 2003
    - 2004
    - 2100
    - 2101
    - 2102
    - 2103
    type: integer
    x-enum-varnames:
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrServiceUnavailable
    - ErrTimeout
    - ErrValidation
    - ErrTenantNotFound
    - ErrTenantAlreadyExists
    - ErrTenantInactive
    - ErrTenantNameRequired
    - ErrTenantInvalidStatus
    - ErrAgentMissingThinkingModel
    - ErrAgentMissingAllowedTools
    - ErrAgentInvalidMaxIterations
    - ErrAgentInvalidTemperature
...
  github_com_Tencent_WeKnora_internal_types.ContextCompressionStrategy:
    enum:
    - sliding_window
    - smart
    type: string
    x-enum-varnames:
    - ContextCompressionSlidingWindow
    - ContextCompressionSmart
  github_com_Tencent_WeKnora_internal_types.ContextConfig:
    properties:
      compression_strategy:
        allOf:
        - $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.ContextCompressionStrategy'
        description: 'Compression strategy: "sliding_window" or "smart"'
      max_tokens:
        description: Maximum tokens allowed in LLM context
        type: integer
      recent_message_count:
        description: |-
          For sliding_window: number of messages to keep
          For smart: number of recent messages to keep uncompressed
        type: integer
      summarize_threshold:
        description: 'Summarize threshold: number of messages before summarization'
        type: integer
    type: object
...
  github_com_Tencent_WeKnora_internal_types.FallbackStrategy:
    enum:
    - fixed
    - model
    type: string
    x-enum-comments:
      FallbackStrategyFixed: Fixed response
      FallbackStrategyModel: Model fallback response
    x-enum-descriptions:
    - Fixed response
    - Model fallback response
    x-enum-varnames:
    - FallbackStrategyFixed
    - FallbackStrategyModel
...
  github_com_Tencent_WeKnora_internal_types.Session:
    properties:
      agent_config:
        allOf:
        - $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.SessionAgentConfig'
        description: Agent 配置（会话级别，仅存储enabled和knowledge_bases）
      context_config:
        allOf:
        - $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.ContextConfig'
        description: 上下文管理配置（可选）
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: Description
        type: string
      embedding_top_k:
        description: 向量召回TopK
        type: integer
      enable_rewrite:
        description: 多轮改写开关
        type: boolean
      fallback_response:
        description: 固定回复内容
        type: string
      fallback_strategy:
        allOf:
        - $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.FallbackStrategy'
        description: 兜底策略
      id:
        description: ID
        type: string
      keyword_threshold:
        description: 关键词召回阈值
        type: number
      knowledge_base_id:
        description: Strategy configuration
        type: string
      max_rounds:
        description: 多轮保持轮数
        type: integer
      rerank_model_id:
        description: 排序模型ID
        type: string
      rerank_threshold:
        description: 排序阈值
        type: number
      rerank_top_k:
        description: 排序TopK
        type: integer
      summary_model_id:
        description: 总结模型ID
        type: string
      summary_parameters:
        allOf:
        - $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.SummaryConfig'
        description: 总结模型参数
      tenant_id:
        description: Tenant ID
        type: integer
      title:
        description: Title
        type: string
      updated_at:
        type: string
      vector_threshold:
        description: 向量召回阈值
        type: number
    type: object
  github_com_Tencent_WeKnora_internal_types.SessionAgentConfig:
    properties:
      agent_mode_enabled:
        description: Whether agent mode is enabled for this session
        type: boolean
      knowledge_bases:
        description: Accessible knowledge base IDs for this session
        items:
          type: string
        type: array
      web_search_enabled:
        description: Whether web search is enabled for this session
        type: boolean
    type: object
...
  github_com_Tencent_WeKnora_internal_types.SummaryConfig:
    properties:
      context_template:
        description: Context template
        type: string
      frequency_penalty:
        description: Frequency penalty
        type: number
      max_completion_tokens:
        description: Max completion tokens
        type: integer
      max_tokens:
        description: Max tokens
        type: integer
      no_match_prefix:
        description: No match prefix
        type: string
      presence_penalty:
        description: Presence penalty
        type: number
      prompt:
        description: Prompt
        type: string
      repeat_penalty:
        description: Repeat penalty
        type: number
      seed:
        description: Seed
        type: integer
      temperature:
        description: Temperature
        type: number
      top_k:
        description: TopK
        type: integer
      top_p:
        description: TopP
        type: number
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handler_session.CreateSessionRequest:
    properties:
      agent_config:
        allOf:
        - $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.SessionAgentConfig'
        description: 'Agent configuration (optional, session-level config only: enabled
          and knowledge_bases)'
      knowledge_base_id:
        description: ID of the associated knowledge base (optional, can be set/changed
          during queries)
        type: string
      session_strategy:
        allOf:
        - $ref: '#/definitions/internal_handler_session.SessionStrategy'
        description: Session strategy configuration
    type: object
...
  internal_handler_session.SessionStrategy:
    properties:
      embedding_top_k:
        description: Number of top results to retrieve from vector search
        type: integer
      enable_rewrite:
        description: Whether to enable query rewrite for multi-round conversations
        type: boolean
      fallback_response:
        description: Fixed response content for fallback
        type: string
      fallback_strategy:
        allOf:
        - $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.FallbackStrategy'
        description: Strategy to use when no relevant knowledge is found
      keyword_threshold:
        description: Threshold for keyword-based retrieval
        type: number
      max_rounds:
        description: Maximum number of conversation rounds to maintain
        type: integer
      no_match_prefix:
        description: Prefix for responses when no match is found
        type: string
      rerank_model_id:
        description: ID of the model used for reranking results
        type: string
      rerank_threshold:
        description: Threshold for reranking results
        type: number
      rerank_top_k:
        description: Number of top results after reranking
        type: integer
      summary_model_id:
        description: ID of the model used for summarization
        type: string
      summary_parameters:
        allOf:
        - $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.SummaryConfig'
        description: Parameters for the summary model
      vector_threshold:
        description: Threshold for vector-based retrieval
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    name: WeKnora Github
    url: https://github.com/Tencent/WeKnora
  description: WeKnora 知识库管理系统 API 文档
  termsOfService: http://swagger.io/terms/
  title: WeKnora API
  version: "1.0"
paths:
...
  /sessions:
    get:
      consumes:
      - application/json
      description: 获取当前租户的会话列表，支持分页
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 会话列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 获取会话列表
      tags:
      - 会话
    post:
      consumes:
      - application/json
      description: 创建新的对话会话
      parameters:
      - description: 会话创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_session.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建的会话
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 创建会话
      tags:
      - 会话
  /sessions/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的会话
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 会话不存在
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 删除会话
      tags:
      - 会话
    get:
      consumes:
      - application/json
      description: 根据ID获取会话详情
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 会话详情
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 会话不存在
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 获取会话详情
      tags:
      - 会话
    put:
      consumes:
      - application/json
      description: 更新会话属性
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: string
      - description: 会话信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.Session'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的会话
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 会话不存在
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 更新会话
      tags:
      - 会话
...
securityDefinitions:
  Bearer:
    description: 输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"